myproject/requirements.txt
certifi==2020.6.20
cycler==0.10.0
Django==2.2.5
django-cleanup==5.0.0
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
